requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
python-jose[full]
passlib[bcrypt]
pytest
pytest-cov
docker